@@ -14,8 +14,11 @@
 		}
 	],
 	"require": {
-		"php": ">=7.4",
-		"typo3/cms-core": "^11.5||^12.4"
+		"php": ">=8.1",
+		"typo3/cms-core": "^12.4"
+	},
+	"require-dev": {
+		"roave/security-advisories": "dev-latest"
 	},
 	"conflict": {
 	},
